third_party/blink/web_tests/http/tests/pointer-lock/requestPointerLock-can-not-transfer-between-documents-expected.txt

Test iframe from same origin can not transfer pointer lock across documents.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

     Lock targetDiv1.
PASS onpointerlockchange received after: Lock targetDiv1.
PASS document.pointerLockElement is targetDiv1
PASS message is "inner-iframe.html onpointerlockerror"
PASS successfullyParsed is true

TEST COMPLETE
